refactor: move publishing diagnostics tests
